Improve the Swift Anagram challenge solution
Some test cases were timing out with the previous solution
because of the usage of a dictionary to map the number of
occurrences of each character.

Moreover, the Character class seems to have a certain overhead
while when using the ascii value of the character as the Array's
index all tests passed.